tried to get rid of the red v over c++ classes thumbnails in the editor

+++ Source/Learn_UE_CPP_101/Learn_UE_CPP_101Character.h
@@ -85,7 +85,7 @@ public:
 	/** Returns FollowCamera subobject **/
 	FORCEINLINE class UCameraComponent* GetFollowCamera() const { return Camera; }
 
-	int32 Health = 100;
+	float Health = 100;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	UCapsuleComponent* Collision;
